Move descriptor and sender factories to new handlers submodule
